move project to point-cloud
