+++ point-cloud/src/main/scala/geotrellis/spark/io/hadoop/formats/PointCloudInputFormat.scala
@@ -1,0 +1,74 @@
+/*
+ * Copyright 2016 Azavea
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package geotrellis.spark.io.hadoop.formats
+
+import geotrellis.spark.pointcloud.json._
+import geotrellis.util.Filesystem
+
+import io.pdal._
+import org.apache.hadoop.fs._
+import org.apache.hadoop.mapreduce._
+import org.apache.hadoop.mapreduce.lib.input._
+
+import java.io.{BufferedOutputStream, File, FileOutputStream}
+
+/** Process files from the path through PDAL, and reads all files point data as an Array[Byte] **/
+class PointCloudInputFormat extends FileInputFormat[Path, Iterator[PointCloud]] {
+  override def isSplitable(context: JobContext, fileName: Path) = false
+
+  override def createRecordReader(split: InputSplit, context: TaskAttemptContext): RecordReader[Path, Iterator[PointCloud]] = {
+    val tmpDir = Filesystem.createDirectory()
+
+    new BinaryFileRecordReader({ bytes =>
+      val remotePath = split.asInstanceOf[FileSplit].getPath
+
+      // copy remote file into local tmp dir
+      val localPath = new File(tmpDir, remotePath.getName)
+      val bos = new BufferedOutputStream(new FileOutputStream(localPath))
+      Stream.continually(bos.write(bytes))
+      bos.close()
+
+      val pipeline = new Pipeline(fileToPipelineJson(localPath).toString)
+
+      AnyRef.synchronized { pipeline.initialise() }
+      AnyRef.synchronized { pipeline.execute }
+
+      val pointViewIterator = AnyRef.synchronized { pipeline.pointViews() }
+      // conversion to list to load everything into JVM memory
+      val packedPoints = pointViewIterator.toList.map { pointView =>
+        val packedPoint = AnyRef.synchronized {
+          pointView.getPointCloud(
+            metadata = AnyRef.synchronized { pipeline.getMetadata() },
+            schema = AnyRef.synchronized { pipeline.getSchema() }
+          )
+        }
+
+        AnyRef.synchronized { pointView.dispose() }
+        packedPoint
+      }.toIterator
+
+      val result = split.asInstanceOf[FileSplit].getPath -> packedPoints
+
+      AnyRef.synchronized { pointViewIterator.dispose() }
+      AnyRef.synchronized { pipeline.dispose() }
+      localPath.delete()
+      tmpDir.delete()
+
+      result
+    })
+  }
+}
