+++ point-cloud/src/test/resources/las/1.2-with-color.las
