+++ point-cloud/src/test/scala/geotrellis/spark/PointsTestEnvironment.scala
