--- points/src/main/scala/geotrellis/spark/points/tiling/Implicits.scala
@@ -1,35 +1,0 @@
-/*
- * Copyright 2016 Azavea
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.spark.points.tiling
-
-import io.pdal._
-import geotrellis.spark._
-import geotrellis.vector._
-import geotrellis.spark.tiling.TilerKeyMethods
-import geotrellis.util._
-import org.apache.spark.rdd._
-
-object Implicits extends Implicits
-
-trait Implicits {
-  implicit class withTilerMethods[K](self: RDD[(K, PackedPoints)]) extends TilerMethods[K](self)
-
-  implicit class withTupleTilerKeyMethods[K: SpatialComponent](val self: (K, Extent)) extends TilerKeyMethods[(K, Extent), K] {
-    def extent = self._2
-    def translate(spatialKey: SpatialKey): K = self._1.setComponent(spatialKey)
-  }
-}
