--- points/src/main/scala/geotrellis/spark/points/tiling/TilerMethods.scala
@@ -1,36 +1,0 @@
-/*
- * Copyright 2016 Azavea
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.spark.points.tiling
-
-import io.pdal._
-import geotrellis.spark._
-import geotrellis.spark.tiling.{LayoutDefinition, TilerKeyMethods}
-import geotrellis.util.MethodExtensions
-
-import org.apache.spark.rdd._
-
-import scala.reflect.ClassTag
-
-class TilerMethods[K](val self: RDD[(K, PackedPoints)]) extends MethodExtensions[RDD[(K, PackedPoints)]] {
-  def tileToLayout[K2: SpatialComponent: ClassTag](layoutDefinition: LayoutDefinition)
-                                                  (implicit ev: K => TilerKeyMethods[K, K2]): RDD[(K2, PackedPoints)] =
-    CutPackedPoints[K, K2](self, layoutDefinition)
-
-  def tileToLayout[K2: SpatialComponent: ClassTag](tileLayerMetadata: TileLayerMetadata[K2])
-                                                  (implicit ev: K => TilerKeyMethods[K, K2]): RDD[(K2, PackedPoints)] =
-    tileToLayout(tileLayerMetadata.layout)
-}
