+++ scripts/buildall.sh
@@ -7,7 +7,7 @@
 ./sbt -J-Xmx2G "project geotools" test || { exit 1; }
 HOSTALIASES=/tmp/hostaliases ./sbt -J-Xmx2G "project geowave" test || { exit 1; }
 ./sbt -J-Xmx2G "project hbase" test  || { exit 1; }
-./sbt -J-Xmx2G "project points" test || { exit 1; }
+./sbt -J-Xmx2G "project point-cloud" test || { exit 1; }
 ./sbt -J-Xmx2G "project proj4" test || { exit 1; }
 ./sbt -J-Xmx2G "project raster-test" test || { exit 1; }
 ./sbt -J-Xmx2G "project raster-testkit" compile || { exit 1; }
