+++ scripts/createHeaders.sh
@@ -8,6 +8,7 @@
       "project geowave" createHeaders test:createHeaders \
       "project hbase" createHeaders  test:createHeaders  \
       "project pdal" createHeaders test:createHeaders \
+      "project point-cloud" createHeaders test:createHeaders \
       "project proj4" createHeaders test:createHeaders \
       "project raster" createHeaders test:createHeaders \
       "project raster-test" createHeaders test:createHeaders \
